feat(app): scatter plot with regression line
